[modify] 댓글 작성

+++ src/main/java/com/spring/boot/reply/repository/ReplyRepositoryCustom.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.spring.boot.entity.Reply;
 
 public interface ReplyRepositoryCustom {
-	List<Reply> list(int board);
+	List<Reply> list(long board);
 }
